Move generated outside of src
