+++ .gitignore
@@ -4,5 +4,5 @@ apps/e2e/node_modules
 apps/e2e/cypress/screenshots
 apps/backend/node_modules
 apps/frontend/src/generated
-apps/backend/src/generated
+apps/backend/generated
 .DS_Store
