+++ .prettierignore
@@ -4,7 +4,7 @@
 **/.github/workflows/
 **/README.md
 **/.vscode/
-**/src/generated/
+**/generated/
 
 # backend
 /apps/backend/src/datasources/stfc/UOWSSoapInterface.ts
