+++ apps/backend/.dockerignore
@@ -6,4 +6,3 @@
 !package*
 !tsconfig.json
 !index.ts
-src/generated
