+++ apps/backend/.gitignore
@@ -5,7 +5,7 @@ coverage
 .DS_Store
 downloads
 uploads
-src/generated
+generated
 
 .idea
